backend/requirements.txt
absl-py==0.8.0
appnope==0.1.0
asn1crypto==0.24.0
astor==0.8.0
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
boto==2.49.0
boto3==1.9.248
botocore==1.12.248
certifi==2018.8.24
cffi==1.11.5
chainer==6.3.0
chardet==3.0.4
codecov==2.0.15
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.4
cryptography==2.3.1
cycler==0.10.0
Cython==0.29.13
decorator==4.4.0
defusedxml==0.6.0
dill==0.3.1.1
Django==2.1.1
django-extensions==2.1.2
django-pandas==0.5.1
django-rest-swagger==2.2.0
djangorestframework==3.8.2
docutils==0.15.2
easydict==1.9
entrypoints==0.3
fasttext==0.9.1
filelock==3.0.12
gast==0.3.1
gensim==3.8.1
google-pasta==0.1.7
graphviz==0.10.1
grpcio==1.23.0
h5py==2.10.0
idna==2.7
imageio==2.4.1
imageio-ffmpeg==0.3.0
imgaug==0.4.0
importlib-metadata==0.23
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itypes==1.1.0
jedi==0.15.1
Jinja2==2.10
jmespath==0.9.4
joblib==0.13.2
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.0
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
llvmlite==0.29.0
Markdown==3.1.1
MarkupSafe==1.0
matplotlib==3.1.1
meson==0.50.0
mistune==0.8.4
more-itertools==7.2.0
mysqlclient==1.3.13
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.4
nltk==3.4.5
notebook==6.0.1
numba==0.45.1
numpy==1.16.4
openapi-codec==1.3.2
opencv-python==4.2.0.34
packaging==19.2
pandas==0.23.4
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
pluggy==0.13.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.7.1
ptyprocess==0.6.0
pwlf==1.1.3
py==1.8.0
pybind11==2.4.2
pycparser==2.19
pyDOE==0.3.8
Pygments==2.4.2
PyMySQL==0.9.2
pyparsing==2.4.2
PyQt5==5.14.1
PyQt5-sip==12.7.1
pyrsistent==0.15.4
pytest==5.1.2
pytest-cov==2.7.1
python-dateutil==2.7.3
pyts==0.9.0
pytz==2018.5
PyWavelets==1.1.1
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.5
requests==2.19.1
s3transfer==0.2.1
saxpy==1.0.1.dev167
scikit-image==0.16.1
scikit-learn==0.21.3
scipy==1.2.1
seaborn==0.9.0
Send2Trash==1.5.0
Shapely==1.7.0
simplejson==3.16.0
six==1.11.0
smart-open==1.8.4
svm==0.1.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
torch==1.3.0
torchfile==0.1.0
torchvision==0.4.1
tornado==6.0.3
traitlets==4.3.2
tslearn==0.2.2
typing==3.6.6
typing-extensions==3.6.6
uritemplate==3.0.0
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.6
widgetsnbextension==3.5.1
word2vec==0.10.2
wrapt==1.11.2
xmltodict==0.12.0
zipp==0.6.0
